rate-limiter-ips/pyproject.toml
[tool.poetry]
name = "rate-limiter-ips"
version = "0.1.0"
description = ""
authors = ["RafaelN <rafaeln_nascimento@hotmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
celery = "^5.5.2"
reddis = "^1.0.1"
fastapi = "^0.115.12"
redis = "^6.1.0"
dotenv = "^0.9.9"
pika = "^1.3.2"
schedule = "^1.2.2"
psycopg2-binary = "^2.9.10"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"